% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_params.R
\name{assign_params}
\alias{assign_params}
\title{evaluate comma separated function parameter in global environment}
\usage{
assign_params()
}
\description{
evaluate comma separated function parameter in global environment
}

